+++ src/coordinates/calculations/index.ts
@@ -1,0 +1,5 @@
+import * as coordinateCalc from './coordinateCalc';
+import * as apparentCoordinateCalc from './apparentCoordinateCalc';
+import * as precessionCalc from './precessionCalc';
+
+export {coordinateCalc, apparentCoordinateCalc, precessionCalc};