+++ src/earth/Earth.ts
@@ -1,11 +1,11 @@
 import AstronomicalObject from '../astronomicalObject/AstronomicalObject';
-import {earthCalc} from '../utils';
 import {EclipticSphericalCoordinates, RectangularCoordinates} from '../coordinates/types/CoordinateTypes';
 import {getAsyncCachedCalculation} from '../cache/calculationCache';
-import {calculateVSOP87, calculateVSOP87Angle} from '../utils/vsop87Calc';
+import {calculateVSOP87, calculateVSOP87Angle} from '../planets/calculations/vsop87Calc';
 import {normalizeAngle} from '../utils/angleCalc';
-import {spherical2rectangular} from '../utils/coordinateCalc';
+import {spherical2rectangular} from '../coordinates/calculations/coordinateCalc';
 import TimeOfInterest from '../time/TimeOfInterest';
+import {earthCalc} from './calculations';
 
 export default class Earth extends AstronomicalObject {
     constructor(toi?: TimeOfInterest) {