+++ src/moon/Moon.ts
@@ -1,14 +1,13 @@
-import {moonCalc, moonPhaseCalc, observationCalc} from '../utils';
+import {observationCalc} from '../utils';
 import AstronomicalObject from '../astronomicalObject/AstronomicalObject';
 import {EclipticSphericalCoordinates, RectangularCoordinates} from '../coordinates/types/CoordinateTypes';
 import TimeOfInterest from '../time/TimeOfInterest';
-import {getApparentMagnitudeMoon} from '../utils/magnitudeCalc';
 import {
     rectangular2spherical,
     rectangularGeocentric2rectangularHeliocentric,
     spherical2rectangular,
-} from '../utils/coordinateCalc';
-import {correctEffectOfNutation} from '../utils/apparentCoordinateCalc';
+} from '../coordinates/calculations/coordinateCalc';
+import {correctEffectOfNutation} from '../coordinates/calculations/apparentCoordinateCalc';
 import Sun from '../sun/Sun';
 import Earth from '../earth/Earth';
 import createSun from '../sun/createSun';
@@ -17,6 +16,8 @@ import {Location} from '../earth/types/LocationTypes';
 import {getRise, getSet, getTransit} from '../utils/riseSetTransitCalc';
 import {createTimeOfInterest} from '../time';
 import {STANDARD_ALTITUDE_MOON_CENTER_REFRACTION} from '../constants/standardAltitude';
+import {getApparentMagnitudeMoon} from './calculations/magnitudeCalc';
+import {moonCalc, moonPhaseCalc} from './calculations';
 import {
     MOON_PHASE_FIRST_QUARTER,
     MOON_PHASE_FULL_MOON,