+++ src/planets/Neptune.ts
@@ -1,9 +1,9 @@
-import {calculateVSOP87, calculateVSOP87Angle} from '../utils/vsop87Calc';
 import {getAsyncCachedCalculation} from '../cache/calculationCache';
 import {EclipticSphericalCoordinates} from '../coordinates/types/CoordinateTypes';
 import {normalizeAngle} from '../utils/angleCalc';
-import {getApparentMagnitudeNeptune} from '../utils/magnitudeCalc';
 import TimeOfInterest from '../time/TimeOfInterest';
+import {getApparentMagnitudeNeptune} from './calculations/magnitudeCalc';
+import {calculateVSOP87, calculateVSOP87Angle} from './calculations/vsop87Calc';
 import {DIAMETER_NEPTUNE} from './constants/diameters';
 import Planet from './Planet';