+++ src/planets/Planet.ts
@@ -2,7 +2,7 @@ import {
     rectangular2spherical,
     rectangularHeliocentric2rectangularGeocentric,
     spherical2rectangular,
-} from '../utils/coordinateCalc';
+} from '../coordinates/calculations/coordinateCalc';
 import AstronomicalObject from '../astronomicalObject/AstronomicalObject';
 import {RectangularCoordinates} from '../coordinates/types/CoordinateTypes';
 import {EclipticSphericalCoordinates} from '../coordinates/types/CoordinateTypes';
@@ -16,7 +16,7 @@ import {
     correctEffectOfAberration,
     correctEffectOfNutation,
     getLightTimeCorrectedJulianDay,
-} from '../utils/apparentCoordinateCalc';
+} from '../coordinates/calculations/apparentCoordinateCalc';
 import {createTimeOfInterest} from '../time';
 import {getRise, getSet, getTransit} from '../utils/riseSetTransitCalc';
 import {Location} from '../earth/types/LocationTypes';