+++ src/solarEclipse/types/besselianElementsTypes.ts
@@ -1,0 +1,19 @@
+export type BesselianElementsTypes = {
+    tMax: number,
+    t0: number,
+
+    dT: number,
+
+    x: Array<number>,
+    y: Array<number>,
+    d: Array<number>,
+    l1: Array<number>,
+    l2: Array<number>,
+    mu: Array<number>,
+
+    tanF1: number,
+    tanF2: number,
+
+    latGreatestEclipse: number,
+    lonGreatestEclipse: number,
+};