+++ src/utils/index.ts
@@ -1,24 +1,25 @@
+import {earthCalc} from '../earth/calculations';
+import {moonCalc} from '../moon/calculations';
+import {moonPhaseCalc} from '../moon/calculations';
+import {sunCalc} from '../sun/calculations';
+import {timeCalc} from '../time/calculations';
+import {starCalc} from '../stars/calculations';
+import {conjunctionCalc} from '../planets/calculations';
+import {coordinateCalc, precessionCalc} from '../coordinates/calculations';
 import * as angleCalc from './angleCalc';
-import * as coordinateCalc from './coordinateCalc';
 import * as distanceCalc from './distanceCalc';
-import * as earthCalc from './earthCalc';
-import * as moonCalc from './moonCalc';
-import * as moonPhaseCalc from './moonPhaseCalc';
 import * as observationCalc from './observationCalc';
-import * as precession from './precessionCalc';
-import * as starCalc from './starCalc';
-import * as sunCalc from './sunCalc';
-import * as timeCalc from './timeCalc';
 
 export {
     angleCalc,
+    conjunctionCalc,
     coordinateCalc,
     distanceCalc,
     earthCalc,
     moonCalc,
     moonPhaseCalc,
     observationCalc,
-    precession,
+    precessionCalc,
     starCalc,
     sunCalc,
     timeCalc,