+++ src/utils/riseSetTransitCalc.ts
@@ -1,14 +1,19 @@
 import {Location} from '../earth/types/LocationTypes';
 import {createTimeOfInterest} from '../time';
 import IAstronomicalObject from '../astronomicalObject/interfaces/IAstronomicalObject';
-import {getDeltaT, getGreenwichApparentSiderealTime, julianDay2julianCenturiesJ2000, julianDay2time} from './timeCalc';
+import {
+    getDeltaT,
+    getGreenwichApparentSiderealTime,
+    julianDay2julianCenturiesJ2000,
+    julianDay2time,
+} from '../time/calculations/timeCalc';
+import {equatorialSpherical2topocentricHorizontalByLocalHourAngle} from '../coordinates/calculations/coordinateCalc';
 import {
     getDeclinationInterpolationArray,
     getRightAscensionInterpolationArray,
     tabularInterpolation3,
 } from './interpolationCalc';
 import {deg2rad, normalizeAngle, rad2deg} from './angleCalc';
-import {equatorialSpherical2topocentricHorizontalByLocalHourAngle} from './coordinateCalc';
 
 export async function getTransit(
     ObjConstructor: any,